Fix eslint and tests
